+++ carp_core/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## 0.20.2
+* improvements to the ` FromJsonFactory().fromJson()` method.
+* added a example of how to support json serialization of polymorphic classes using `Serialization`.
+
+## 0.20.2
 * `DataPoint` updated to support both up- and download to the CARP server
 
 ## 0.20.1